move NodeJS/global module declarations to module.d.ts
- fix module.require (does not include namespace)
- deprecate global NodeModule, NodeRequire
