+++ types/node/test/globals.ts
@@ -10,16 +10,6 @@
 
 declare var RANDOM_GLOBAL_VARIABLE: true;
 
-// global aliases for compatibility
-{
-    const x: NodeModule = {} as any;
-    const y: NodeModule = {} as any;
-    x.children.push(y);
-    x.parent = require.main!;
-    require.main = y;
-    x.path; // $ExpectType string
-}
-
 // exposed gc
 {
     if (gc) {
