+++ types/node/test/module.ts
@@ -17,8 +17,8 @@ let rf: (m: string) => any;
 rf = Module.createRequire("mod");
 rf = Module.createRequire(new URL("file:///C:/path/"));
 
-const aModule: NodeModule = new Module("s");
-const bModule: NodeModule = new Module("b", aModule);
+const aModule: NodeJS.Module = new Module("s");
+const bModule: NodeJS.Module = new Module("b", aModule);
 
 const builtIn: string[] = Module.builtinModules;
 const builtinResult: boolean = Module.isBuiltin("node:fs");
