+++ types/node/v18/globals.d.ts
@@ -106,24 +106,10 @@ declare global {
     *                                               *
     ------------------------------------------------*/
 
-    // For backwards compability
-    interface NodeRequire extends NodeJS.Require {}
-    interface RequireResolve extends NodeJS.RequireResolve {}
-    interface NodeModule extends NodeJS.Module {}
-
     var global: typeof globalThis;
 
     var process: NodeJS.Process;
     var console: Console;
-
-    var __filename: string;
-    var __dirname: string;
-
-    var require: NodeRequire;
-    var module: NodeModule;
-
-    // Same as module.exports
-    var exports: any;
 
     interface GCFunction {
         (options: {
@@ -327,49 +313,6 @@ declare global {
         interface RefCounted {
             ref(): this;
             unref(): this;
-        }
-
-        interface Require {
-            (id: string): any;
-            resolve: RequireResolve;
-            cache: Dict<NodeModule>;
-            /**
-             * @deprecated
-             */
-            extensions: RequireExtensions;
-            main: Module | undefined;
-        }
-
-        interface RequireResolve {
-            (id: string, options?: { paths?: string[] | undefined }): string;
-            paths(request: string): string[] | null;
-        }
-
-        interface RequireExtensions extends Dict<(m: Module, filename: string) => any> {
-            ".js": (m: Module, filename: string) => any;
-            ".json": (m: Module, filename: string) => any;
-            ".node": (m: Module, filename: string) => any;
-        }
-        interface Module {
-            /**
-             * `true` if the module is running during the Node.js preload
-             */
-            isPreloading: boolean;
-            exports: any;
-            require: Require;
-            id: string;
-            filename: string;
-            loaded: boolean;
-            /** @deprecated since v14.6.0 Please use `require.main` and `module.children` instead. */
-            parent: Module | null | undefined;
-            children: Module[];
-            /**
-             * @since v11.14.0
-             *
-             * The directory name of the module. This is usually the same as the path.dirname() of the module.id.
-             */
-            path: string;
-            paths: string[];
         }
 
         interface Dict<T> {
